url_helpers: Switch some int's to size_t's
Switch the return values for the strlen()'s in match_uri() from int's to
size_t's to match strlen(3)'s definition.

+++ www/url_helpers.c
@@ -1347,8 +1347,8 @@ void send_template(const char *template, TMPL_varlist *varlist,
  */
 bool match_uri(const char *request_uri, const char *match)
 {
-	int rlen;
-	int mlen = strlen(match);
+	size_t rlen;
+	size_t mlen = strlen(match);
 	const char *request;
 	char *req = strdupa(request_uri);
